Add theme button and use it in account and _components
